Converted VSUnit tests to XUnit.  Tested FHIR calls.
