+++ src/EhrgoHealth.WebService.UnitTests/packages.config
@@ -1,4 +1,13 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Newtonsoft.Json" version="8.0.2" targetFramework="net452" />
+  <package id="Hl7.Fhir.DSTU21" version="0.90.4" targetFramework="net452" />
+  <package id="Newtonsoft.Json" version="6.0.7" targetFramework="net452" />
+  <package id="xunit" version="2.1.0" targetFramework="net452" />
+  <package id="xunit.abstractions" version="2.0.0" targetFramework="net452" />
+  <package id="xunit.assert" version="2.1.0" targetFramework="net452" />
+  <package id="xunit.core" version="2.1.0" targetFramework="net452" />
+  <package id="xunit.extensibility.core" version="2.1.0" targetFramework="net452" />
+  <package id="xunit.extensibility.execution" version="2.1.0" targetFramework="net452" />
+  <package id="xunit.runner.console" version="2.2.0-beta1-build3239" targetFramework="net452" developmentDependency="true" />
+  <package id="xunit.runner.visualstudio" version="2.1.0" targetFramework="net452" />
 </packages>
